% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculFournees.R
\name{calculFournees}
\alias{calculFournees}
\title{Calcul fournees}
\usage{
calculFournees(pains, Vm = 60)
}
\arguments{
\item{pains}{df. Table des pains.}

\item{Vm}{int. Volume maximum d'une fournee.}
}
\description{
Calcul fournees
}
