% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculPainGUI.R
\name{calculPainGUI}
\alias{calculPainGUI}
\title{Calcul Pain Global User Interface}
\usage{
calculPainGUI()
}
\value{
An html user interface.
}
\description{
Calcul Pain Global User Interface
}
\examples{
\dontrun{
 calculPainGUI()
}

}
