% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPoids.R
\name{genPoids}
\alias{genPoids}
\title{Generate Poids}
\usage{
genPoids(min = 0, max = 300, mean = 90, sd = 20)
}
\arguments{
\item{min}{int. Minmal value.}

\item{max}{int. Maximal value.}

\item{mean}{int. Mean value for normal distribution.}

\item{sd}{int. SD value for normal distribution.}
}
\description{
Generate Poids
}
