% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFournees.R
\name{graphFournees}
\alias{graphFournees}
\title{Graphique des fournees}
\usage{
graphFournees(pains, fournees)
}
\arguments{
\item{pains}{df. Table des pains.}

\item{fournees}{df. Table des fournees.}
}
\description{
Graphique des fournees
}
