% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFournees.R
\name{graphFournees}
\alias{graphFournees}
\title{Graphique des fournees}
\usage{
graphFournees(fournees)
}
\arguments{
\item{fournees}{df. Table des fournees.}
}
\value{

}
\description{
Graphique des fournees
}
